Time: 0 ms (100.00%) | Memory: 95.6 MB (33.90%) - LeetSync

+++ 4-median-of-two-sorted-arrays/median-of-two-sorted-arrays.cpp
@@ -4,9 +4,9 @@ public:
         nums1.insert(nums1.end(), nums2.begin(), nums2.end());
         sort(nums1.begin(), nums1.end());
         int n =nums1.size();
-        double first =0;
-        double last = nums1.size() - 1;
-        double mid = (first + last)/2;
+        int first =0;
+        int last = nums1.size() - 1;
+        int mid = (first + last)/2;
         if(n%2 == 0){
           return (nums1[mid]+nums1[mid+1])/2.00;
